Restructure filesystem
Moved the src files into their own directory.
